Move `ContactRole` related codes to `contact_role.go` file
